Update community.json
This demo enables a Fable app that can be used instead of the Vue.js app that is in Microsoft Serverless SignalR Demo.  This may be helpful to people getting started and trying to see how they could plug into SignalR.  This is just meant to be a minmal starting point and keeps the same conventions as the Vue.js app for intercompatibility.